Functionality - Transfer money. (5) 1: transferDto -> transferDto.getMoney() in BankAccountService.transferMoney. 2: Money - Integer -> BigDecimal

+++ src/main/java/com/xandr/pep_aston/dto/BankAccountDto.java
@@ -5,11 +5,13 @@ import jakarta.validation.constraints.Positive;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class BankAccountDto {
 
-    private Integer money;
+    private BigDecimal money;
 
     @NotBlank
     private String name;
